requirements.txt
streamlit
deepface
scikit-learn
pillow
numpy